fix test

+++ test/access-log.js
@@ -3,18 +3,24 @@ const cfntest = require('@cfn-modules/test');
 
 test.serial('access-log', async t => {
   const stackName = cfntest.stackName();
+  let accessLogBucketName = null;
   try {
     t.log(await cfntest.createStack(`${__dirname}/access-log.yml`, stackName, {}));
     const outputs = await cfntest.getStackOutputs(stackName);
     t.log(outputs);
     // what could we test here?
-    t.log(await cfntest.emptyBucket(outputs.AccessLogBucketName));
+    accessLogBucketName = outputs.AccessLogBucketName;
+    t.log(await cfntest.emptyBucket(accessLogBucketName));
   } finally {
     try {
       t.log(await cfntest.deleteStack(stackName));
     } catch (err) {
-      t.log(await cfntest.emptyBucket(outputs.AccessLogBucketName));
-      t.log(await cfntest.deleteStack(stackName));
+      if (accessLogBucketName != null) {
+        t.log(await cfntest.emptyBucket(outputs.AccessLogBucketName));
+        t.log(await cfntest.deleteStack(stackName));
+      } else {
+        throw err;
+      }
     }
     t.pass();
   }
